Merzbild.jl internal API reference

The public API functions are exported by the module

Particle indexing

Merzbild.map_cont_index — Function

map a continuous index in [0, n_local-1] to an index in the particle array given particle_indexer struct describing the split

map a continuous index in [0, n_local-1] to an index in the particle array given particle_indexer struct describing the split

Merzbild.update_particle_indexer_new_lower_count — Function

Update particle indexer when we reduced the particle count

Merzbild.update_particle_indexer_new_particle — Function

Update particle indexer when we add a new particle

Base.getindex — Function

Get underlying particle in ParticleVector with index i

Base.setindex! — Function

Set underlying particle in ParticleVector with index i to a new particle

Base.length — Function

Get length of a ParticleVector` instance

Base.resize! — Function

Resize a ParticleVector` instance

Merzbild.update_particle_buffer_new_particle — Method

Update the buffer of unused Particles in a ParticleVector instance

Merzbild.update_particle_buffer_new_particle — Method

Update the buffer of unused Particles in a ParticleVector instance

Merzbild.update_particle_buffer_new_particle — Method

Dummy function in case Vector{Particle} is used and not a ParticleVector`

Merzbild.update_particle_buffer_new_particle — Method

Dummy function in case Vector{Particle} is used and not a ParticleVector`

Loading species and interaction data

Merzbild.find_species_in_db — Function

Find a chemical species in an LXCat-format XML file

Merzbild.load_elastic_data — Function

Load elastic electron-neutral scattering data from an LXCat-format XML file

Merzbild.load_ionization_data — Function

Load electron-impact ionization data from an LXCat-format XML file

Merzbild.compute_mu_ref — Function

Compute reference viscosity

Sampling

Merzbild.create_unit_dvgrid — Function

generate a grid with extent [-1.0,1.0]x[-1.0,1.0]x[-1.0,1.0]

Merzbild.create_noiseless_dvgrid — Function

generate a grid [-vxmax, vxmax]x[-vymax, vymax]x[-vzmax, vzmax]

Merzbild.create_vdf — Function

create an empty VDF of size nx * ny * nz

Merzbild.sample_bkw! — Function

Sample from the BKW distribution with temperature T for a species with mass m at t=0 and add a velocity offset

Sample from the BKW distribution with temperature T for a species with mass m at t=0 and add a velocity offset

Merzbild.evaluate_distribution_on_grid! — Function

Evaluate a distribution on a discrete velocity grid

Merzbild.sample_maxwellian_single! — Function

Sample a single particle from a Maxwell distribution with temperature T for a species with mass m and add a velocity offset

Merzbild.sample_maxwellian! — Function

Sample nparticles particles from a Maxwell distribution with temperature T for a species with mass m and add a velocity offset

Sample nparticles particles from a Maxwell distribution with temperature T for a species with mass m and add a velocity offset

Collision computations

Merzbild.compute_n_coll_single_species — Function

Compute number of collisions, collisions between particles of same species

Merzbild.compute_n_coll_two_species — Function

Compute number of collisions, collisions between particles of different species

Merzbild.compute_vhs_factor — Function

Compute interaction-specific factor $\pi D_{VHS}^2 (2 T_{ref,VHS}/m_r)^{(\omega_{VHS} - 0.5)} \frac{1}{\Gamma(2.5 - \omega_{VHS})}$

Merzbild.compute_com! — Function

Compute center of mass velocity

Merzbild.compute_g! — Function

Compute relative velocity

Merzbild.compute_g_new_ionization! — Function

Compute post-ionization relative velocity

Merzbild.scatter_vhs! — Function

Scatter two particles using VHS scattering

Merzbild.scatter_electron_vhs! — Function

Scatter an electron off of a neutral particle using VHS scattering

Merzbild.scatter_ionization_electrons! — Function

Scatter two electrons

Merzbild.sigma_vhs — Function

Compute the VHS cross-section for a given interaction

Merzbild.compute_tabulated_cs_constant_continuation — Function

Compute a cross-section from tabulated data, continuing with first and last values if outside of range

Merzbild.compute_tabulated_cs_zero_continuation — Function

Compute a cross-section from tabulated data, continuing with 0.0 if outside of range

Merzbild.compute_cross_sections_only! — Function

Compute electron-impact ionization and excitation cross-sections, and electron-neutral elastic scattering Return collision energy in eV

Merzbild.compute_cross_sections! — Function

Compute electron-impact ionization and excitation cross-sections, and electron-neutral elastic scattering Plus compute total cross-section and probabilites of different processes

Merzbild.get_cs_total — Function

Get the value of the total collision cross-section for electron-neutral interactions

Merzbild.get_cs_elastic — Function

Get the value of the elastic scattering electron-neutral cross-section

Merzbild.get_cs_ionization — Function

Get the value of the electron-impact ionization cross-section

Merzbild.get_ionization_threshold — Function

Get the ionization threshold

Merzbild.get_electron_energy_split — Function

Get the way electron energy is split during ionization

Merging

Merzbild.GridCell — Type

Struct for keeping track of merging-related quantities in a velocity grid cell

Merzbild.compute_velocity_extent! — Function

Compute extent of velocity grid based on temperature in the cell

Merzbild.compute_grid_index — Function

Compute index of particle on the merging grid

Merzbild.clear_merging_grid! — Function

Reset merging grid

Merzbild.compute_grid! — Function

Compute properties for all cells on the merging grid

Merzbild.compute_new_particles! — Function

Compute new particles based on the grid cell properties without checking particle locations

Compute post-merge particles

Merzbild.vx_sign — Function

Return sign of velocity vx of an octant in velocity space

Merzbild.vy_sign — Function

Return sign of velocity vy of an octant in velocity space

Merzbild.vz_sign — Function

Return sign of velocity vz of an octant in velocity space

Merzbild.check_speed_bound — Function

Check speed bound

Check speed bound with multiplier

Merzbild.base_multi_index_moments — Function

Base multi indices corresponding to conservation of mass momentum and energy components

Merzbild.compute_w_total_v0! — Function

Compute total computational weight of particles and mean velocity, as well as velocity bounds

Merzbild.ccm — Function

Compute unweighted centered moment

Compute unweighted centered moment

Merzbild.compute_lhs_and_rhs! — Function

Compute LHS matrix and RHS vector for NNLS merging

Merzbild.compute_lhs_and_rhs_rate_preserving! — Function

Compute LHS matrix and RHS vector for NNLS merging, rate-preserving version

Merzbild.compute_lhs_particles_additional! — Function

Compute additional LHS columns for additional particles

Merzbild.compute_lhs_particles_additional_rate_preserving! — Function

Compute additional LHS columns for additional particles, rate preserving

Merzbild.scale_lhs_rhs! — Function

Scale LHS and RHS

Merzbild.scale_lhs_rhs_rate_preserving! — Function

Scale LHS and RHS, rate-preserving version

Merzbild.compute_post_merge_particles_nnls! — Function

Compute post-merge particles

Merzbild.OctreeCell — Type

Struct holding data needed for refinement of an octree bin

Merzbild.OctreeFullCell — Type

Octree bin properties required to merge the particles in a bin

Merzbild.fill_bins — Function

Fill the octree bins struct with zero data

Merzbild.fill_full_bins — Function

Fill the octree full bins struct with zero data

Merzbild.clear_octree! — Function

Reset octree

Merzbild.resize_octree_buffers! — Function

Resize octree buffers

Merzbild.compute_octant — Function

Compute octant of particle velocity relative to a v_middle`

Merzbild.bin_bounds_inherit! — Function

Compute new bin bounds of sub-bin inheriting bounds of parent bin

Merzbild.bin_bounds_recompute! — Function

Recompute bin bounds based on particle velocities

Merzbild.compute_v_mean! — Function

Compute mean velocity of bin

Merzbild.compute_v_median! — Function

Compute median velocity of bin

Merzbild.get_new_bin_id — Function

Get index of a newly created bin

Merzbild.split_bin! — Function

Sort particles into sub-bins, split bin, keeping track of which sub-bins particles end up in

Merzbild.compute_bin_props! — Function

Compute properties in a bin required for merging

Merzbild.get_bin_post_merge_np — Function

Get number of post-merge particles in a bin

Merzbild.init_octree! — Function

Initialize the top bin in an octree

Merzbild.merge_octree_N2_based! — Function

Perform N:2 merging

Grids

Merzbild.Cell1D — Type

Cell element of a 1-D grid

Merzbild.get_cell — Function

Find cell in 1-D uniform grid given coordinates of a particle

Particle-surface interactions

Merzbild.specular_reflection_x! — Function

Perform specular reflection of a particle in the x direction

Merzbild.diffuse_reflection_x! — Function

Perform diffuse reflection of a particle in the x direction

Merzbild.reflect_particle_x! — Function

Reflect particle in the x direction

Constants

Merzbild.c_light — Constant

Speed of light, m/s

Merzbild.eV — Constant

Electron-Volt, K

Merzbild.eV_J — Constant

Electron-Volt, J

Merzbild.eV_J_inv — Constant

1.0/eV[J], 1/J (or equivalent to how much 1 J is equal to expressed in eV)

Merzbild.twopi — Constant

$2 \pi$

Merzbild.e_mass_div_electron_volt — Constant

Electron mass divided by 1 eV, kg/J

Merzbild.direction_signs — Constant

Positive and negative direction signs

Merzbild.q_e — Constant

Elementary charge, C

Misc

Merzbild.compute_thermal_velocity — Function

compute thermal velocity: $\sqrt(2kT/m)$

Merzbild.binary_search — Function

Binary search for value val in array x

Merzbild.linear_interpolation — Function

Perform linear interpolation between two neighbouring values in y

Merzbild.compute_mixed_moment — Function

Compute mixed moment of particles in a cell

NNLS

Merzbild.solve! — Function

Algorithm NNLS: NONNEGATIVE LEAST SQUARES

The original version of this code was developed by Charles L. Lawson and Richard J. Hanson at Jet Propulsion Laboratory 1973 JUN 15, and published in the book "SOLVING LEAST SQUARES PROBLEMS", Prentice-HalL, 1974. Revised FEB 1995 to accompany reprinting of the book by SIAM.

GIVEN AN M BY N MATRIX, A, AND AN M-VECTOR, B, COMPUTE AN N-VECTOR, X, THAT SOLVES THE LEAST SQUARES PROBLEM A * X = B SUBJECT TO X .GE. 0

Merzbild.construct_householder! — Function

CONSTRUCTION AND/OR APPLICATION OF A SINGLE HOUSEHOLDER TRANSFORMATION.. Q = I + U*(U**T)/B

The original version of this code was developed by Charles L. Lawson and Richard J. Hanson at Jet Propulsion Laboratory 1973 JUN 12, and published in the book "SOLVING LEAST SQUARES PROBLEMS", Prentice-HalL, 1974. Revised FEB 1995 to accompany reprinting of the book by SIAM.

Merzbild.fastview — Function

UnsafeVectorView only works for isbitstype types. For other types, we're already allocating lots of memory elsewhere, so creating a new View is fine.

This function looks type-unstable, but the isbitstype(T) test can be evaluated by the compiler, so the result is actually type-stable.

Fallback for non-contiguous arrays, for which UnsafeVectorView does not make sense.

Merzbild.solve_triangular_system! — Function

The original version of this code was developed by Charles L. Lawson and Richard J. Hanson at Jet Propulsion Laboratory 1973 JUN 15, and published in the book "SOLVING LEAST SQUARES PROBLEMS", Prentice-HalL, 1974. Revised FEB 1995 to accompany reprinting of the book by SIAM.

Merzbild.UnsafeVectorView — Type

Views in Julia still allocate some memory (since they need to keep a reference to the original array). This type allocates no memory and does no bounds checking. Use it with caution.

Merzbild.orthogonal_rotmat — Function

COMPUTE ORTHOGONAL ROTATION MATRIX.. The original version of this code was developed by Charles L. Lawson and Richard J. Hanson at Jet Propulsion Laboratory 1973 JUN 12, and published in the book "SOLVING LEAST SQUARES PROBLEMS", Prentice-HalL, 1974. Revised FEB 1995 to accompany reprinting of the book by SIAM.

COMPUTE.. MATRIX (C, S) SO THAT (C, S)(A) = (SQRT(A2+B2)) (-S,C) (-S,C)(B) ( 0 ) COMPUTE SIG = SQRT(A2+B2) SIG IS COMPUTED LAST TO ALLOW FOR THE POSSIBILITY THAT SIG MAY BE IN THE SAME LOCATION AS A OR B .

Merzbild.apply_householder! — Function

CONSTRUCTION AND/OR APPLICATION OF A SINGLE HOUSEHOLDER TRANSFORMATION.. Q = I + U*(U**T)/B

The original version of this code was developed by Charles L. Lawson and Richard J. Hanson at Jet Propulsion Laboratory 1973 JUN 12, and published in the book "SOLVING LEAST SQUARES PROBLEMS", Prentice-HalL, 1974. Revised FEB 1995 to accompany reprinting of the book by SIAM.