Merzbild.jl public API reference

Particles

Merzbild.Particle — Type

The Particle struct

Merzbild.ParticleVector — Type

The ParticleVector struct

Particle indexing

Merzbild.ParticleIndexer — Type

The ParticleIndexer struct

Merzbild.ParticleIndexerArray — Type

The ParticleIndexerArray struct

Loading species and interaction data

Merzbild.Species — Type

The Species struct

Merzbild.Interaction — Type

Used to store interaction parameters for a 2-species interaction

Merzbild.load_species_data — Function

Load a vector of species data

Load a vector of species data for a single species

Merzbild.load_interaction_data — Function

Load interaction data

Load interaction data filling with dummy data if needed

Merzbild.load_interaction_data_with_dummy — Function

Load interaction data filling with dummy data if needed

Merzbild.load_species_and_interaction_data — Function

Load species and interaction data filling with dummy data if needed

Merzbild.load_electron_neutral_interactions — Function

Load electron-neutral interaction data from an LXCat-format XML file

Sampling

Merzbild.maxwellian — Function

Evaluate a Maxwell distribution with temperature T for a species with mass m at a velocity vx vy vz

Merzbild.bkw — Function

Evaluate the BKW distribution with temperature T for a species with mass m at a velocity vx vy vz at a scaled time t

Merzbild.sample_on_grid! — Function

Sample from an arbitrary distribution on a discrete velocity grid

Merzbild.sample_maxwellian_on_grid! — Function

Sample from a Maxwellian on a discrete velocity grid

Merzbild.sample_particles_equal_weight! — Function

Sample from a distribution for particles of a specific species in a specific cell

Sample particles in each cell of 1-D uniform grid given ppc

Sample particles in each cell of 1-D uniform grid given target number density

Computing macroscopic properties

Merzbild.PhysProps — Type

Structure to store computed physical properties in a cell

Merzbild.PhysProps — Method

PhysProps(n_cells, n_species, moments_list; Tref=300.0)

Construct physical properties given a number of cells and species

Merzbild.PhysProps — Method

PhysProps(pia, moments_list; Tref=300.0)

Construct physical properties given a ParticleIndexerArray instance

Merzbild.PhysProps — Method

PhysProps(pia)

Construct physical properties given a ParticleIndexerArray instance, no moments computed

Merzbild.compute_props! — Function

Compute PhysProps

Merzbild.compute_props_with_total_moments! — Function

Compute PhysProps with total moments

Merzbild.compute_props_sorted! — Function

Compute physical properties of all species in all cells assuming the particles are sorted; moments are not computed

Merzbild.avg_props! — Function

Average PhysProps

Merzbild.clear_props! — Function

Clear all data from PhysProps, for use when physical properties are averaged over timesteps and averaging over a new set of timesteps needs to be started

Collision computations

Merzbild.CollisionData — Type

Used to store temporary collision data for a specific collision

Merzbild.CollisionData — Method

Create empty CollisionData instance

Merzbild.CollisionFactors — Type

Struct to store NTC-related collision factors

Merzbild.CollisionFactors — Method

Create a CollisionFactors instance

Merzbild.CollisionDataFP — Type

Used to store Fokker-Planck collision data

Merzbild.CollisionDataFP — Method

Create empty CollisionDataFP instance

Merzbild.create_collision_factors_array — Method

Create collision factors for all interaction pairs for a 0-D case (1 spatial cell)

Merzbild.create_collision_factors_array — Method

Create collision factors for all interaction pairs for a non 0-D case

Merzbild.create_computed_crosssections — Function

Create a vector of ComputedCrossSections instances

Merzbild.estimate_sigma_g_w_max — Function

Estimate $(\sigma g w)_{max}$ for a two-species interaction

Estimate $(\sigma g w)_{max}$ for a single-species interaction

Merzbild.estimate_sigma_g_w_max! — Function

Estimate $(\sigma g w)_{max}$ for all species

Merzbild.estimate_sigma_g_w_max_ntc_n_e! — Function

Estimate $(\sigma g w)_{max}$ for an electron-neutral interaction by stochastically choosing particle pairs

Merzbild.ntc! — Method

Perform elastic collisions between particles of same species using the NTC algorithm

Merzbild.ntc! — Method

Perform elastic collisions between particles of different species using the NTC algorithm

Merzbild.ntc_n_e! — Function

Perform elastic scattering and ionizing collisions for electron-neutral interactions

Merzbild.ntc_n_e_es! — Function

Perform elastic scattering and ionizing collisions for electron-neutral interactions using the event splitting approach of Oblapenko et al. (2022)

Merzbild.fp! — Function

Model single-species elastic collisions using a linear Fokker-Planck approximation

Merging

Grid merging

Merzbild.GridN2Merge — Type

Struct for merging on a velocity grid

Fields:

Nx: number of cells

Merzbild.GridN2Merge — Method

GridN2Merge(Nx::Int, Ny::Int, Nz::Int, extent_multiplier::T) where T <: AbstractArray

Create velocity grid-based merging

Positional arguments:

Nx: number of cells in vx direction

Ny: number of cells in vy direction

Merzbild.GridN2Merge — Method

GridN2Merge(N::Int, extent_multiplier::T) where T <: AbstractArray

Create velocity grid-based merging with equal number of cells in each direction

Merzbild.GridN2Merge — Method

GridN2Merge(Nx::Int, Ny::Int, Nz::Int, extent_multiplier::Float64)

Create velocity grid-based merging with single multiplier

Merzbild.GridN2Merge — Method

GridN2Merge(Nx::Int, Ny::Int, Nz::Int,
            extent_multiplier_x::Float64,
            extent_multiplier_y::Float64,
            extent_multiplier_z::Float64)

Create velocity grid-based merging

Merzbild.GridN2Merge — Method

GridN2Merge(N::Int, extent_multiplier::Float64)

Create velocity grid-based merging with equal number of cells in each direction, single multiplier

Merzbild.merge_grid_based! — Function

Merge particles using a velocity grid-based merging approach

NNLS merging

Merzbild.NNLSMerge — Type

Struct for keeping track of merging-related quantities for NNLS merging

Merzbild.NNLSMerge — Method

NNLSMerge(multi_index_moments, init_np; rate_preserving=false)

Create NNLS-based merging

Positional arguments:

multi_index_moments: vector of mixed moments to preserve, [(i1, j1, k1), (i2, j2, k2), ...] Mass/momentum/directionalenergy are always conserved! (Added automatically if not in list)

init_np: initial number of particles to pre-allocated memory for

Keyword arguments:

rate_preserving: used for rate-preserving merging of electrons, preserves approximate elastic collision and ionization rates

Merzbild.compute_multi_index_moments — Function

Compute all mixed moments of total order up to n

Merzbild.merge_nnls_based! — Function

Perform NNLS-based merging

Merzbild.merge_nnls_based_rate_preserving! — Function

Perform NNLS-based merging, rate-preserving version

Octree merging

Merzbild.OctreeBinSplit — Type

OctreeBinSplit OctreeBinMidSplit=1 OctreeBinMeanSplit=2 OctreeBinMedianSplit=3

Enum for octree bin splitting types

Merzbild.OctreeInitBin — Type

OctreeInitBin OctreeInitBinMinMaxVel=1 OctreeInitBinMinMaxVelSym=2 OctreeInitBinC=3

Enum for initial bin bounds

Merzbild.OctreeBinBounds — Type

OctreeBinBounds OctreeBinBoundsInherit=1 OctreeBinBoundsRecompute=2

Computation of bounds of split bin

Merzbild.OctreeN2Merge — Type

Struct for N:2 Octree merging

Merzbild.OctreeN2Merge — Method

OctreeN2Merge(split::OctreeBinSplit; init_bin_bounds=OctreeInitBinMinMaxVel, bin_bounds_compute=OctreeBinBoundsInherit,
          max_Nbins=4096, max_depth=10)

Create Octree N:2 merging

Grids and particle sorting

Merzbild.Grid1DUniform — Type

1-D Uniform grid

Merzbild.GridSortInPlace — Type

Struct for in-place sorting of particles

Merzbild.sort_particles! — Function

Sort particles on a grid

Particle movement

Merzbild.convect_particles! — Function

Convect particles on 1-D uniform grid

Particle-surface interactions

Merzbild.MaxwellWallBC — Type

A struct to hold information about a diffuse reflecting wall

Merzbild.MaxwellWalls1D — Type

A struct to hold information about all diffuse reflecting walls in the simulation

Merzbild.MaxwellWalls1D — Method

Create a MaxwellWalls struct for a 1-D simulation with 2 walls ("left" and "right)

I/O

Merzbild.write_grid — Function

Write info to a NetCDF file

Merzbild.IOSkipList — Type

Struct that holds track of which variables are not to be written to NetCDF file

Merzbild.IOSkipList — Method

IOSkipList(list_of_variables_to_skip)

Construct an IOSkipList from a list of variable names

Merzbild.IOSkipList — Method

IOSkipList()

Construct an empty IOSkipList

Merzbild.NCDataHolder — Type

Struct that holds NetCDF-output related data

Merzbild.NCDataHolder — Method

NCDataHolder(nc_filename, names_skip_list, species_data, phys_props; global_attributes=Dict{Any,Any}())

Construct a NCDataHolder instance

Merzbild.NCDataHolder — Method

NCDataHolder(nc_filename, species_data, phys_props; global_attributes=Dict{Any,Any}())

Construct a NCDataHolder instance with an empty list of variable to skip

Merzbild.close_netcdf — Function

Close NetCDF file

Merzbild.write_netcdf_phys_props — Function

Write PhysProps to NetCDF file

Particle-in-Cell

Merzbild.accelerate_constant_field_x! — Function

Accelerate particles with a constant electric field in the X direction

Constants

Merzbild.k_B — Constant

Boltzmann constant, J/K